+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/ChatClient.kt
@@ -72,6 +72,7 @@ import io.getstream.chat.android.client.events.UserEvent
 import io.getstream.chat.android.client.extensions.ATTACHMENT_TYPE_FILE
 import io.getstream.chat.android.client.extensions.ATTACHMENT_TYPE_IMAGE
 import io.getstream.chat.android.client.extensions.cidToTypeAndId
+import io.getstream.chat.android.client.extensions.internal.isLaterThanDays
 import io.getstream.chat.android.client.extensions.retry
 import io.getstream.chat.android.client.header.VersionPrefixHeader
 import io.getstream.chat.android.client.helpers.AppSettingManager
@@ -108,6 +109,9 @@ import io.getstream.chat.android.client.notifications.PushNotificationReceivedLi
 import io.getstream.chat.android.client.notifications.handler.NotificationConfig
 import io.getstream.chat.android.client.notifications.handler.NotificationHandler
 import io.getstream.chat.android.client.notifications.handler.NotificationHandlerFactory
+import io.getstream.chat.android.client.persistance.repository.RepositoryFacade
+import io.getstream.chat.android.client.persistance.repository.factory.RepositoryFactory
+import io.getstream.chat.android.client.persistance.repository.noop.NoOpRepositoryFactory
 import io.getstream.chat.android.client.plugin.Plugin
 import io.getstream.chat.android.client.plugin.factory.PluginFactory
 import io.getstream.chat.android.client.plugin.listeners.ChannelMarkReadListener
@@ -167,6 +171,7 @@ import okhttp3.OkHttpClient
 import java.io.File
 import java.util.Calendar
 import java.util.Date
+import kotlin.time.Duration.Companion.days
 import io.getstream.chat.android.client.experimental.socket.ChatSocket as ChatSocketExperimental
 
 /**
@@ -191,7 +196,9 @@ internal constructor(
     private val initializationCoordinator: InitializationCoordinator = InitializationCoordinator.getOrCreate(),
     private val appSettingsManager: AppSettingManager,
     private val chatSocketExperimental: ChatSocketExperimental,
+    private val pluginFactories: List<PluginFactory>,
     lifecycle: Lifecycle,
+    private val repositoryFactoryProvider: RepositoryFactory.Provider,
 ) {
     private val logger = StreamLog.getLogger("Chat:Client")
     private val waitConnection = MutableSharedFlow<Result<ConnectionData>>()
@@ -205,6 +212,18 @@ internal constructor(
             }
         }
     )
+
+    @InternalStreamChatApi
+    public val repositoryFacade: RepositoryFacade
+        get() = _repositoryFacade
+            ?: (getCurrentUser() ?: getStoredUser())
+                ?.let {
+                    createRepositoryFacade(createReposotiryFactory(it))
+                        .also { _repositoryFacade = it }
+                }
+            ?: createRepositoryFacade()
+
+    private var _repositoryFacade: RepositoryFacade? = null
 
     /**
      * With clientState allows the user to get the state of the SDK like connection, initialization...
@@ -298,10 +317,6 @@ internal constructor(
         }
     }
 
-    internal fun addPlugins(plugins: List<Plugin>) {
-        this.plugins = plugins
-    }
-
     @InternalStreamChatApi
     public fun addInterceptor(interceptor: Interceptor) {
         this.interceptors.add(interceptor)
@@ -370,9 +385,8 @@ internal constructor(
             }
             userState is UserState.NotSet -> {
                 logger.v { "[setUser] user is NotSet" }
-                initializationCoordinator.userConnectionRequest(user)
+                initializeClientWithUser(user, cacheableTokenProvider, isAnonymous)
                 clientState.toMutableState()?.setUser(user)
-                initializeClientWithUser(cacheableTokenProvider, isAnonymous)
                 userStateService.onSetUser(user, isAnonymous)
                 if (ToggleService.isSocketExperimental()) {
                     chatSocketExperimental.connectUser(user, isAnonymous)
@@ -420,15 +434,24 @@ internal constructor(
     }
 
     private fun initializeClientWithUser(
+        user: User,
         tokenProvider: CacheableTokenProvider,
         isAnonymous: Boolean,
     ) {
+        _repositoryFacade = createRepositoryFacade(createReposotiryFactory(user))
+        plugins = pluginFactories.map { it.get(user) }
         // fire a handler here that the chatDomain and chatUI can use
         config.isAnonymous = isAnonymous
         tokenManager.setTokenProvider(tokenProvider)
         appSettingsManager.loadAppSettings()
         warmUp()
     }
+
+    private fun createReposotiryFactory(user: User): RepositoryFactory =
+        repositoryFactoryProvider.createRepositoryFactory(user)
+
+    private fun createRepositoryFacade(repositoFactory: RepositoryFactory = NoOpRepositoryFactory): RepositoryFacade =
+        RepositoryFacade.create(repositoFactory, scope)
 
     /**
      * Get the current settings of the app. Check [AppSettings].
@@ -510,9 +533,8 @@ internal constructor(
         }
 
         userCredentialStorage.get()?.let { config ->
-            initializationCoordinator.userConnectionRequest(User(id = config.userId).apply { name = config.userName })
-
             initializeClientWithUser(
+                User(id = config.userId).apply { name = config.userName },
                 tokenProvider = CacheableTokenProvider(ConstantTokenProvider(config.userToken)),
                 isAnonymous = config.isAnonymous,
             )
@@ -1040,10 +1062,12 @@ internal constructor(
                 chatSocketExperimental.disconnect(DisconnectCause.ConnectionReleased)
             }
             if (flushPersistence) {
+                repositoryFacade.clear()
                 userCredentialStorage.clear()
             }
             lifecycleObserver.dispose()
             appSettingsManager.clear()
+            _repositoryFacade = null
             Result.success(Unit)
         }
 
@@ -2270,8 +2294,8 @@ internal constructor(
      * Returns all events that happened for a list of channels since last sync (while the user was not
      * connected to the web-socket).
      *
-     * @param channelsIds The list of channel CIDs
-     * @param lastSyncAt The last time the user was online and in sync
+     * @param channelsIds The list of channel CIDs. Cannot be empty.
+     * @param lastSyncAt The last time the user was online and in sync. Shouldn't be later than 30 days.
      *
      * @return Executable async [Call] responsible for obtaining missing events.
      */
@@ -2282,11 +2306,27 @@ internal constructor(
     ): Call<List<ChatEvent>> {
         return api.getSyncHistory(channelsIds, lastSyncAt)
             .withPrecondition(scope) {
-                when (channelsIds.isEmpty()) {
-                    true -> Result.error(ChatError("channelsIds must contain at least 1 id."))
-                    else -> Result.success(Unit)
-                }
+                checkSyncHistoryPreconditions(channelsIds, lastSyncAt)
             }
+    }
+
+    /**
+     * Checks if sync history request parameters meet preconditions:
+     * 1. If [channelsIds] is not empty.
+     * 2. If [lastSyncAt] is no later than 30 days
+     */
+    private fun checkSyncHistoryPreconditions(channelsIds: List<String>, lastSyncAt: Date): Result<Unit> {
+        return when {
+            channelsIds.isEmpty() -> {
+                Result.error(ChatError("channelsIds must contain at least 1 id."))
+            }
+            lastSyncAt.isLaterThanDays(THIRTY_DAYS_IN_MILLISECONDS) -> {
+                Result.error(ChatError("lastSyncAt cannot by later than 30 days."))
+            }
+            else -> {
+                Result.success(Unit)
+            }
+        }
     }
 
     /**
@@ -2429,6 +2469,7 @@ internal constructor(
         private var retryPolicy: RetryPolicy = NoRetryPolicy()
         private var distinctApiCalls: Boolean = true
         private var debugRequests: Boolean = false
+        private var repositoryFactoryProvider: RepositoryFactory.Provider? = null
 
         /**
          * Sets the log level to be used by the client.
@@ -2552,6 +2593,13 @@ internal constructor(
         }
 
         /**
+         * Inject a [RepositoryFactory.Provider] to use your own DB Persistence mechanism.
+         */
+        public fun withRepositoryFactoryProvider(provider: RepositoryFactory.Provider): Builder = apply {
+            repositoryFactoryProvider = provider
+        }
+
+        /**
          * Adds a plugin factory to be used by the client.
          * @see [PluginFactory]
          *
@@ -2595,16 +2643,6 @@ internal constructor(
             this.distinctApiCalls = false
         }
 
-        private fun configureInitializer(chatClient: ChatClient) {
-            chatClient.initializationCoordinator.addUserConnectionRequestListener { user ->
-                chatClient.addPlugins(
-                    pluginFactories.map { pluginFactory ->
-                        pluginFactory.get(user)
-                    }
-                )
-            }
-        }
-
         public override fun build(): ChatClient {
             return super.build()
         }
@@ -2615,6 +2653,7 @@ internal constructor(
             replaceWith = ReplaceWith("this.build()"),
             level = DeprecationLevel.ERROR
         )
+        @SuppressWarnings("LongMethod")
         override fun internalBuild(): ChatClient {
             if (apiKey.isEmpty()) {
                 throw IllegalStateException("apiKey is not defined in " + this::class.java.simpleName)
@@ -2677,10 +2716,14 @@ internal constructor(
                 retryPolicy = retryPolicy,
                 appSettingsManager = appSettingsManager,
                 chatSocketExperimental = module.experimentalSocket(),
-                lifecycle = lifecycle
-            ).also {
-                configureInitializer(it)
-            }
+                lifecycle = lifecycle,
+                pluginFactories = pluginFactories,
+                repositoryFactoryProvider = repositoryFactoryProvider
+                    ?: pluginFactories
+                        .filterIsInstance<RepositoryFactory.Provider>()
+                        .firstOrNull()
+                    ?: NoOpRepositoryFactory.Provider,
+            )
         }
 
         private fun setupStreamLog() {
@@ -2738,6 +2781,7 @@ internal constructor(
         private const val KEY_MESSAGE_ACTION = "image_action"
         private const val MESSAGE_ACTION_SEND = "send"
         private const val MESSAGE_ACTION_SHUFFLE = "shuffle"
+        private val THIRTY_DAYS_IN_MILLISECONDS = 30.days.inWholeMilliseconds
 
         private const val ARG_TYPING_PARENT_ID = "parent_id"