+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/persistance/repository/QueryChannelsRepository.kt
@@ -40,4 +40,9 @@ public interface QueryChannelsRepository {
      * @param querySort [QuerySorter]
      */
     public suspend fun selectBy(filter: FilterObject, querySort: QuerySorter<Channel>): QueryChannelsSpec?
+
+    /**
+     * Clear QueryChannels of this repository.
+     */
+    public suspend fun clear()
 }