+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/offline/plugin/listener/internal/QueryMembersListenerImpl.kt
@@ -20,9 +20,9 @@ import io.getstream.chat.android.client.api.models.FilterObject
 import io.getstream.chat.android.client.api.models.querysort.QuerySorter
 import io.getstream.chat.android.client.extensions.internal.toCid
 import io.getstream.chat.android.client.models.Member
+import io.getstream.chat.android.client.persistance.repository.RepositoryFacade
 import io.getstream.chat.android.client.plugin.listeners.QueryMembersListener
 import io.getstream.chat.android.client.utils.Result
-import io.getstream.chat.android.offline.repository.builder.internal.RepositoryFacade
 
 /**
  * [QueryMembersListener] implementation for [io.getstream.chat.android.offline.plugin.internal.OfflinePlugin].